Move RECIPE entries from the "fluid" prototype folder into "recipes"
Currently fluids and recipes for fluids are mixed together under
the "fluid" prototype folder. This is part of a much wider problem
that causes load order issues on several levels and likely leads
to quite a few hacks in data-updates.lua -
Which we really should try to avoid, as it makes understanding and
extending the code base much harder (in my case, simply trying to
update icons for a couple of recipes).
We already have the structures that we need to make things
better established, so all we need to do is adhere to them :)

This patch cleans up the fluid prototype section by moving all recipes
there into its own file in the "recipes" folder.
These changes shouldn't effect existing mod behaviour.