Move FLUID entry prototypes into a single file
Moving the recipes out of the fluid prototypes
had the effect that we have now a whole bunch of
files who each contain only the definition for
a singular fluid.

This is annoying to navigate, maintain and extend,
so we move everything into a singular file
to make everyones life a bit easier.

Does not change mod behaviour.