Move ITEM entries from the "buildings" prototype folder into "items"
The "buildings" prototype folder has the same problem as "fluids" had,
expect worse. Items, recipes and entities are mixed together here.
This is part of a much wider problem that causes load order issues on
several levels and likely leads to quite a few hacks in data-updates.lua -
We already have the structures that we need to make things
better established, so all we need to do is adhere to them :)

This patch cleans up the building prototype section by moving
all items there into its own file in the "items" folder.
These changes shouldn't effect existing mod behaviour.

Attention: The item for "pulp-mill" was left in its place due
to startup problems. It looks like the whole item/recipe/entity
shouldn't exist anymore. I'm unsure how to proceed here,
so i left it as it is for now.