Load RECIPE prototypes after all other prototypes
Currently it is impossible to predict when which
ITEM or FLUID is loaded, which makes it quite
difficult to refer to these prototypes when
using py.composite_icon for example.

This does not solve all problems, as a lot of RECIPE
prototypes are currently not in the recipes prototype folder,
but it will ease that pain a little bit.